feat: move cross-corr funcs into shared file and improve compilation stability
